requirements.txt
Pillow~=9.1.0
PyMySQL~=1.0.2
pyTelegramBotAPI~=4.4.0
pymysql-pool~=0.3.7
python-dotenv~=0.20.0
